test(desktop): keep prompt race coverage within lint budgets
Problem:
The initial prompt-race patch added coverage to three files that already exceeded the repository size ceiling, causing repolint to reject their incremental growth.

Root cause:
The scoped replay and A-X-A regression cases were placed beside broad legacy suites, and the small replay bridge helper remained inside the oversized controller module.

Fix:
Move the Go scoped replay test and frontend replay helper into focused files, and relocate the A-X-A prompt race into the existing stale-navigation suite without changing production behavior.

Verification:
go run ./tools/repolint
go test ./... -count=1
cd desktop && go test . -count=1
cd desktop && go test -race . -run Test(RuntimeReattachFencesPendingAskBeforeReplay|ReplayPendingPromptsForTabDoesNotReplaySiblingAsk)$ -count=1
cd desktop/frontend && npx tsc --noEmit
cd desktop/frontend && npx tsc -p tsconfig.test.json --noEmit
cd desktop/frontend && npx eslint src/lib/useController.ts src/lib/promptReplay.ts src/__tests__/activate-topic-stale.test.tsx src/__tests__/tab-switch-hydration.test.tsx
